fix refactor home
