+++ components/home/ExploreCard.jsx
@@ -1,6 +1,6 @@
 import { motion } from "framer-motion";
 import Link from "next/link";
-import fadeIn from "../lib/framer/animations/fadeIn";
+import fadeIn from "../../lib/framer/animations/fadeIn";
 
 const ExploreCard = ({
   id,
